wip: transient hlmatch proof-of-concept
No configurability in here, just seeing if this could work. It does.
However, the hlmatch is scoped per window, which means if you switch
windows right after it's been applied, the timer will then fail to clear
the match highlight in the original window. It's a bit of an edge case,
but could probably be handled by an autocommand on the WinLeave event.