very_magic_slash: return early if LoupeVeryMagic=0
It's useful to be able to toggle g:LoupeVeryMagic on/off for different
types of searches. The `/`, `?` commands generate a very-magic string
dynamically, yielding the expected behavior, but `:` command-line
behavior is based on the value g:LoupeVeryMagic had when the plugin was
loaded.

Returning early from very_magic_slash if g:LoupeVeryMagic is off yields
the expected behavior.